backend/requirements.txt
flask
flask-cors
language-tool-python
textblob
google-genai
python-dotenv
titlecase
bcrypt
PyJWT
nltk
textstat
